apps/python-sdk/requirements.txt
requests
httpx
pytest
pytest-asyncio
python-dotenv
websockets
nest-asyncio
pydantic>=2.0
aiohttp
